% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_binary.R
\name{count_binary}
\alias{count_binary}
\title{Formatted count and percentage of 1's}
\usage{
count_binary(x, digits = 0)
}
\arguments{
\item{x}{A binary/boolean vector.}

\item{digits}{Number of digits used when formatting percentage.}
}
\value{
Length 1 character of the form \code{"n (pct\%)"} with \code{n} and \code{pct}
replaced with number of occurences and percentage of 1's in \code{x}.
}
\description{
Formatted count and percentage of 1's
}
\examples{
x <- rnorm(100)
count_binary(x > 1)
}
