% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_elementwise.R
\name{format_elementwise}
\alias{format_elementwise}
\title{Apply \code{format} elementwise}
\usage{
format_elementwise(x, digits, scientific = FALSE, ...)
}
\arguments{
\item{x}{A vector.}

\item{digits}{Number of digits after the decimal place.}

\item{scientific}{Boolean. as in \code{\link[=format]{format()}}.}

\item{...}{Arguments passed to \code{\link[=format]{format()}}.}
}
\value{
\code{x} after applying format elementwise.
}
\description{
Apply \code{format} elementwise
}
\examples{
x <- c(11, 1, 1.1, 1.11, 1.111111111)
format_elementwise(x, 1)
}
