% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{ggtheme}
\alias{ggtheme}
\title{Change active \code{ggplot2} theme and default color palettes}
\usage{
ggtheme()
}
\value{
\code{NULL}
}
\description{
Change active \code{ggplot2} theme and default color palettes
}
\examples{
ggtheme()
}
