% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Formatted mean and standard deviation}
\usage{
mean_sd(
  x,
  digits = 1,
  digits.mean = NULL,
  digits.sd = NULL,
  NA.show = TRUE,
  NA.name = "Missing",
  digits.NA = 0
)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits, digits.mean, digits.sd}{Number of digits to print. \code{digits}
applies to both mean and sd, with the other options overriding if supplied.}

\item{NA.show}{Boolean. Whether to show count of missing values.}

\item{NA.name}{Name of row with count of missing values.}

\item{digits.NA}{Decimal digits for percentage of \code{NA}.}
}
\value{
If \code{NA.show} is \code{FALSE} then a length 1 character of the format
\code{"mean, (sd)"} with the mean and sd of \code{x}.

If \code{NA.show} is \code{TRUE} then a data.table with two rows, the first row has \code{NA} in
the first column and the mean and sd in the second column. The second row
has \code{NA.name} in the first column and the count and pct of missing values in
the second.
}
\description{
Formatted mean and standard deviation
}
\examples{
mean_sd(c(rnorm(100), NA))
}
