% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_quantiles.R
\name{median_quantiles}
\alias{median_quantiles}
\title{Formatted median and quantiles}
\usage{
median_quantiles(
  x,
  q = c(0.25, 0.75),
  q.sep = ", ",
  digits = 1,
  digits.median = NULL,
  digits.quantiles = NULL,
  NA.show = TRUE,
  NA.name = "Missing",
  digits.NA = 0
)
}
\arguments{
\item{x}{A numeric vector.}

\item{q}{A length 2 numeric vector of probabilities for the quantiles.}

\item{q.sep}{A length 1 character to separate the quantiles.}

\item{digits, digits.median, digits.quantiles}{Number of digits to print. \code{digits}
applies to both median and quantiles, with the other options overriding if supplied.}

\item{NA.show}{Boolean. Whether to show count of missing values.}

\item{NA.name}{Name of row with count of missing values.}

\item{digits.NA}{Decimal digits for percentage of \code{NA}.}
}
\value{
If \code{NA.show} is \code{FALSE} then a length 1 character of the format
\code{"median, (q[1]q.sepq[2])"} with the median and given quantiles of \code{x}.

If \code{NA.show} is \code{TRUE} then a data.table with two rows, the first row has \code{NA} in
the first column and the median and quantiles in the second column. The second row
has \code{NA.name} in the first column and the count and pct of missing values in
the second.
}
\description{
Formatted median and quantiles
}
\examples{
median_quantiles(c(rnorm(100), NA))
}
