% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_percent.R
\name{n_percent}
\alias{n_percent}
\title{Formatted count and percent of levels}
\usage{
n_percent(x, digits = 0, NA.show = TRUE, NA.name = "Missing")
}
\arguments{
\item{x}{A vector, usually factor or character.}

\item{digits}{Number of digits to use when formatting percentages.}

\item{NA.show}{Boolean. Whether to show count of missing values.}

\item{NA.name}{Name of row with count of missing values.}
}
\value{
A data.table with a column containing names of levels
and a column containing characters of the format \code{"n (pct\%)"} where
\code{n} is the count of the level and \code{pct} is the percentage.
}
\description{
Formatted count and percent of levels
}
\examples{
x <- c("A", "A", "B", "C")
n_percent(x)
}
