% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table_docx.R
\name{save_table_docx}
\alias{save_table_docx}
\title{Save flextable as .docx}
\usage{
save_table_docx(
  table,
  path,
  orientation = "portrait",
  fontname = "Times New Roman",
  fontsize = 11
)
}
\arguments{
\item{table}{A flextable object.}

\item{path}{A character containing the file destination.}

\item{orientation}{A character, either \code{"portrait"} or \code{"landscape"}.}

\item{fontname}{Character. Name of font to use (must be installed on system).}

\item{fontsize}{Numeric. Fontsize.}
}
\value{
Path to saved file.
}
\description{
Save flextable as .docx
}
\examples{
# Load the diamonds data
library(data.table)
library(ggplot2)

data(diamonds)

tab <- summarytable(
  diamonds,
  `Cut, n (\%)` = cut,
  `Weight, mean (sd)` = carat
)
# save_table_docx(tab, "path/to/file/table.docx")
}
\seealso{
\code{\link[=summarytable]{summarytable()}}
}
