% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytable.R
\name{summarytable}
\alias{summarytable}
\title{Create a summary table}
\usage{
summarytable(
  data,
  ...,
  .by = NULL,
  flatten.by = FALSE,
  flatten.by.sep = ", ",
  .by.NA.drop = FALSE,
  .by.NA.name = "Missing",
  N.row = .N,
  N.row.header = TRUE,
  N.row.name = "n",
  default.numeric = median_quantiles,
  default.factor = n_percent,
  total.column = FALSE,
  total.name = "Total",
  variable.name = "",
  indent.char = "  ",
  color.cells = "lightgrey",
  border.inner.v = FALSE,
  border.inner.h = TRUE,
  autofit = TRUE,
  return.list = FALSE
)
}
\arguments{
\item{data}{A data.frame-like object.}

\item{...}{Name-expression pairs \code{LHS = RHS} where LHS is the text to write in the
leftmost column of the table, and RHS is an R expression that will be evaluated in \code{data},
see details and examples.}

\item{.by}{A character vector of column/variable names in \code{data} that the table should
stratify on, i.e. create one column per level of \code{.by} (or combination of levels if multiple
columns are given).}

\item{flatten.by}{Boolean. If \code{FALSE} (default), the table header will have one row for
each variable in \code{.by}. If TRUE, will condense these into a single row.}

\item{flatten.by.sep}{Character of length 1. If \code{flatten.by} is \code{TRUE}, this character
will separate levels of variables in \code{.by} in the header.}

\item{.by.NA.drop, .by.NA.name}{If there are \code{NA} values in the\code{.by} columns, a separate column
will be created for rows with missing values when \code{.by.NA.drop = FALSE}. The name of this column
is controlled by \code{.by.NA.name}}

\item{N.row}{An expression that gives the number observations in each group, which by
default is put in the header.}

\item{N.row.header}{Boolean. If TRUE (default), the result of \code{N.row} is put in the
header. Otherwise it is put as the first row of the table.}

\item{N.row.name}{Character of length 1. Name of \code{N.row} row, or the symbol put into
the header in front of the result.}

\item{default.numeric, default.factor}{A function that will be applied to a numeric or
factor/character variable in data when only its name is given as the RHS of an expression in \code{...}.}

\item{total.column}{Boolean. Only has effect if \code{.by} is not \code{NULL}, in which case
it controls if an column is added for the unstratified data (\code{TRUE}) or not (\code{FALSE}, the default).}

\item{total.name}{Character of length 1. Name of the total column (if present).}

\item{variable.name}{Character of length 1. The name of the leftmost column of the table
(the empty character \code{""} by default).}

\item{indent.char}{Character of length 1. Character to indent sublevels of table by.
Default is two spaces (\code{"  "}), use \code{"\\t"} for a tab character, although this seems
to give rather ugly results when saving to .docx format.}

\item{color.cells}{Character. Color to apply to cells of alternating subtables.
Set \code{NULL} or \code{FALSE} for no coloring.}

\item{border.inner.v, border.inner.h}{Whether to add inner vertical and horizontal rules.}

\item{autofit}{Whether automatically resize the table. If \code{TRUE}, newlines will be ignored.}

\item{return.list}{Boolean. If \code{TRUE}, the function will return a list data.tables
instead of a formatted flextable.}
}
\value{
A formatted flextable object.
}
\description{
A function for easy creation of publication-ready summary tables, i.e. "table 1" in many
medical publications. The function has simple defaults and allows for user-defined summary functions.
}
\details{
The function allows for evaluating arbitrary user-supplied code as the RHS of \code{...}.
However, the result of such an expression must evaluate to either:
\itemize{
\item A scalar (like e.g. the \code{mean()} function).
\item Equivalent to above, a \code{data.table} with one row and two columns, the first
column being \code{NA} and the second containing the scalar value.
\item More generally, a data.table with multiple rows and columns. The last column contains
the values to display in the table, and the remaining columns contain information to identify
the cells, e.g. the levels of a factor. These columns should get finer going from left to right.
If the top row has \code{NA} identifier then the result of that row will not be indented in the final table.
}
}
\examples{
# Load the diamonds data
library(data.table)
library(ggplot2)

data(diamonds)
setDT(diamonds)

diamonds[sample(.N, 10), cut := NA][sample(.N, 10), carat := NA]

# When only giving variables as RHS, the functions default.numeric or
# default.factor are applied as appropriate.
summarytable(
  diamonds,
  `Cut, n (\%)` = cut,
  `Weight, mean (sd)` = carat
)

# Other functions can be applied to columns freely as if they were variables
summarytable(
  diamonds,
  `Cut, n (\%)` = cut,
  `Weight, mean (sd)` = carat,
  `Square root depth` = mean(sqrt(depth)),
  `Price in USD, median (Q1, Q3)` = median_quantiles(price)
)

# We can stratify by discrete variables using .by
summarytable(
  diamonds,
  `Cut, n (\%)` = cut,
  `Weight, mean (sd)` = carat,
  `Square root depth` = mean(sqrt(depth)),
  `Price in USD, median (Q1, Q3)` = median_quantiles(price),
  .by = "color"
)

# We can add a total column to a stratified table
summarytable(
  diamonds,
  `Cut, n (\%)` = cut,
  `Weight, mean (sd)` = carat,
  `Square root depth, mean` = mean(sqrt(depth)),
  `Price in USD, median (Q1, Q3)` = median_quantiles(price),
  .by = "color",
  total.column = TRUE
)
}
\seealso{
\code{\link[=save_table_docx]{save_table_docx()}}
}
