% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarytable_worker.R
\name{summarytable_worker}
\alias{summarytable_worker}
\title{Helper for summarytable}
\usage{
summarytable_worker(
  data,
  ex,
  ex_name,
  .by = NULL,
  .by.NA.drop = FALSE,
  .by.NA.name = "Missing",
  default.numeric = mean_sd,
  default.factor = n_percent,
  flatten.by.sep = ", ",
  total.column = is.null(.by),
  total.name = "Total",
  indent.char = "  "
)
}
\arguments{
\item{data, .by, default.numeric, default.factor, flatten.by.sep, total.column, total.name, indent.char}{See \code{summarytable()}.}

\item{ex}{An expression to evaluate in \code{data}. The RHS of an argument of \code{...}
in a call to summarytable.}

\item{ex_name}{Character of length 1. The LHS of an argument of \code{...} in
a call to summarytable.}
}
\value{
A data.table.
}
\description{
Helper for summarytable
}
