% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_elementwise.R
\name{format_elementwise}
\alias{format_elementwise}
\title{Round numbers to a certain number of significant decimals.}
\usage{
format_elementwise(x, digits, scientific = FALSE, ...)
}
\arguments{
\item{x}{A vector.}

\item{digits}{Number of significant digits after the decimal place.}

\item{scientific}{Boolean. as in \code{\link[=format]{format()}}.}

\item{...}{Arguments passed to \code{\link[=format]{format()}}.}
}
\value{
\code{x} as a character rounded to \code{digits} significant decimal places.
}
\description{
Round numbers to a certain number of significant decimals.
}
\examples{
x <- c(11, 1, 1.1, 1.11, 1.111111111)
format_elementwise(x, 1)
}
