% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_percent.R
\name{n_percent}
\alias{n_percent}
\title{Formatted count and percent of levels}
\usage{
n_percent(
  x,
  digits = 0,
  NA.action = "category",
  NA.report.complete = "Complete, n (\%)",
  NA.report.complete.name = "Categories, n (\%)",
  NA.name = "Missing"
)
}
\arguments{
\item{x}{A vector, usually factor or character.}

\item{digits}{Number of digits to use when formatting percentages.}

\item{NA.action}{Character. How to deal with missing values in \code{x}.
Option \code{"category"} treats \code{NA} as its own level and reports the count. In that case,
option \code{"NA.name"} controls the name of this category.
Option \code{"omit"} simply disregards all \code{NA} values. Percentages will be of non-missing entries.
Option \code{"report.complete"} is like \code{"omit"}, except it reports the number and percentage of complete observations.}

\item{NA.name}{Name of row with count of missing values.}
}
\value{
A data.table with a column containing names of levels
and a column containing characters of the format \code{"n (pct\%)"} where
\code{n} is the count of the level and \code{pct} is the percentage.
}
\description{
Formatted count and percent of levels
}
\examples{
x <- c("A", "A", "B", "C")
n_percent(x)
}
